Convert to int64

+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/Entry.java
@@ -153,7 +153,7 @@ public abstract class Entry<S extends Entry, T extends Version> implements Compa
     @Column
     @ApiModelProperty(value = "Implementation specific timestamp for last updated on webservice. "
             + "Tools-> For automated builds: last time tool/namespace was refreshed Dockstore, tool info (like changing dockerfile path) updated, or default version selected. For hosted tools: when you created the tool. "
-            + "Workflows-> For remote: When refresh all is hit for first time. Hosted: Seems to be time created.", position = 10, dataType = "integer")
+            + "Workflows-> For remote: When refresh all is hit for first time. Hosted: Seems to be time created.", position = 10, dataType = "long")
     private Date lastUpdated;
 
     @Column
@@ -175,12 +175,12 @@ public abstract class Entry<S extends Entry, T extends Version> implements Compa
     // database timestamps
     @Column(updatable = false, nullable = false)
     @CreationTimestamp
-    @ApiModelProperty(dataType = "integer")
+    @ApiModelProperty(dataType = "long")
     private Timestamp dbCreateDate;
 
     @Column(nullable = false)
     @UpdateTimestamp
-    @ApiModelProperty(dataType = "integer")
+    @ApiModelProperty(dataType = "long")
     private Timestamp dbUpdateDate;
 
     @Column
